update cache folder name, move to generics
