+++ lib/msal-common/src/unifiedCache/entities/AppMetadataEntity.ts
@@ -3,12 +3,11 @@
  * Licensed under the MIT License.
  */
 import { APP_META_DATA, Separators } from "../../utils/Constants";
-import { Serializable } from "../serialize/Serializable";
 
 /**
  * APP_META_DATA Cache
  */
-export class AppMetadataEntity extends Serializable {
+export class AppMetadataEntity {
     clientId: string;
     environment: string;
     familyId?: string;
