+++ lib/msal-common/src/unifiedCache/entities/Credential.ts
@@ -3,12 +3,10 @@
  * Licensed under the MIT License.
  */
 
-import { Serializable } from "../serialize/Serializable";
-
 /**
  * Base type for credentials to be stored in the cache: eg: ACCESS_TOKEN, ID_TOKEN etc
  */
-export class Credential extends Serializable {
+export class Credential {
     homeAccountId: string;
     environment: string;
     credentialType: string;
