+++ lib/msal-common/src/unifiedCache/serialize/Deserializer.ts
@@ -32,7 +32,7 @@ export class Deserializer {
                 AccessTokenCacheMaps.fromCacheMap
             );
             const accessToken: AccessTokenEntity = new AccessTokenEntity();
-            accessToken.toObject(mappedAT);
+            CacheHelper.toObject(accessToken, mappedAT);
             atObjects[key] = accessToken;
         });
 
@@ -51,7 +51,7 @@ export class Deserializer {
                 IdTokenCacheMaps.fromCacheMap
             );
             const idToken: IdTokenEntity = new IdTokenEntity();
-            idToken.toObject(mappedIdT);
+            CacheHelper.toObject(idToken, mappedIdT);
             idObjects[key] = idToken;
         });
 
@@ -70,7 +70,7 @@ export class Deserializer {
                 RefreshTokenCacheMaps.fromCacheMap
             );
             const refreshToken: RefreshTokenEntity = new RefreshTokenEntity();
-            refreshToken.toObject(mappedRT);
+            CacheHelper.toObject(refreshToken, mappedRT);
             rtObjects[key] = refreshToken;
         });
 
@@ -88,7 +88,7 @@ export class Deserializer {
                 AccountCacheMaps.fromCacheMap
             );
             const account: AccountEntity = new AccountEntity();
-            account.toObject(mappedAcc);
+            CacheHelper.toObject(account, mappedAcc);
             accountObjects[key] = account;
         });
 
@@ -107,7 +107,7 @@ export class Deserializer {
                 AppMetadataCacheMaps.fromCacheMap
             );
             const amd: AppMetadataEntity = new AppMetadataEntity();
-            amd.toObject(mappedAmd);
+            CacheHelper.toObject(amd, mappedAmd);
             appMetadataObjects[key] = amd;
         });
