+++ lib/msal-common/src/unifiedCache/utils/CacheHelper.ts
@@ -3,7 +3,20 @@
  * Licensed under the MIT License.
  */
 
+import { StringDict } from "../../utils/MsalTypes";
+
 export class CacheHelper {
+    /**
+     * Helper to convert serialized data to object
+     * @param obj
+     * @param json
+     */
+    static toObject<T>(obj: T, json: StringDict): T {
+        for (const propertyName in json) {
+            obj[propertyName] = json[propertyName];
+        }
+        return obj;
+    }
     /**
      * helper function to swap keys and objects
      * @param cacheMap
