+++ lib/msal-common/test/unifiedCache/entities/AccessTokenEntity.spec.ts
@@ -1,5 +1,5 @@
 import { expect } from "chai";
-import { AccessTokenEntity } from "../../../src/uCache/entities/AccessTokenEntity";
+import { AccessTokenEntity } from "../../../src/unifiedCache/entities/AccessTokenEntity";
 import { mockCache } from "./cacheConstants";
 
 describe("AccessTokenEntity.ts Unit Tests", () => {
