+++ lib/msal-common/test/unifiedCache/serialize/Serializer.spec.ts
@@ -1,7 +1,7 @@
 import { expect } from "chai";
-import { CacheInterface } from "../../../src/uCache/serialize/CacheInterface";
+import { CacheInterface } from "../../../src/unifiedCache/serialize/CacheInterface";
 import { mockCache } from "../entities/cacheConstants";
-import { Serializer } from "../../../src/uCache/serialize/Serializer";
+import { Serializer } from "../../../src/unifiedCache/serialize/Serializer";
 
 const cachedJson = require("./cache.json");
